+++ internal_testsuites/typescript/src/testsuite/startosis_module_test/startosis_valid_module_with_no_type.test.ts
@@ -3,7 +3,7 @@ import log from "loglevel";
 import {createEnclave} from "../../test_helpers/enclave_setup";
 import * as path from "path";
 import {DEFAULT_DRY_RUN, EMPTY_EXECUTE_PARAMS, IS_PARTITIONING_ENABLED, JEST_TIMEOUT_MS} from "./shared_constants";
-import {generateScriptOutput} from "../../test_helpers/startosis_helpers";
+import {generateScriptOutput, readStreamContentUntilClosed} from "../../test_helpers/startosis_helpers";
 
 const VALID_MODULE_NO_TYPE_TEST_NAME = "valid-module-no-type";
 const VALID_MODULE_NO_TYPE_REL_PATH = "../../../../startosis/valid-kurtosis-module-no-type"
@@ -26,21 +26,21 @@ test("Test valid startosis module with no type", async () => {
 
         log.info(`Loading module at path '${moduleRootPath}'`)
 
-        const executeStartosisModuleResult = await enclaveContext.executeStartosisModule(moduleRootPath, EMPTY_EXECUTE_PARAMS, DEFAULT_DRY_RUN)
+        const outputStream = await enclaveContext.executeKurtosisModule(moduleRootPath, EMPTY_EXECUTE_PARAMS, DEFAULT_DRY_RUN)
 
-        if (executeStartosisModuleResult.isErr()) {
+        if (outputStream.isErr()) {
             log.error(`An error occurred execute startosis module '${moduleRootPath}'`);
-            throw executeStartosisModuleResult.error
+            throw outputStream.error
         }
-        const executeStartosisModuleValue = executeStartosisModuleResult.value;
+        const [interpretationError, validationErrors, executionError, instructions] = await readStreamContentUntilClosed(outputStream.value);
 
         const expectedScriptOutput = "Hello World!\n"
 
-        expect(generateScriptOutput(executeStartosisModuleValue.getKurtosisInstructionsList())).toEqual(expectedScriptOutput)
+        expect(generateScriptOutput(instructions)).toEqual(expectedScriptOutput)
 
-        expect(executeStartosisModuleValue.getInterpretationError()).toBeUndefined()
-        expect(executeStartosisModuleValue.getExecutionError()).toBeUndefined()
-        expect(executeStartosisModuleValue.getValidationErrors()).toBeUndefined()
+        expect(interpretationError).toBeUndefined()
+        expect(validationErrors).toEqual([])
+        expect(executionError).toBeUndefined()
     } finally {
         stopEnclaveFunction()
     }