+++ docs/changelog.md
@@ -1,9 +1,42 @@
 # TBD
 
-# 0.53.8
+# 0.54.0
+
+### Breaking Changes
+- Renamed `kurtosis exec` to `kurtosis run` and `main in main.star` to `run in main.star`
+  - Upgrade to the latest CLI, and use the `run` function instead
+  - Upgrade existing modules to have `run` and not `main` in `main.star`
+
+### Features
+- Updated the CLI to consume the streaming endpoints to execute Startosis. Kurtosis Instructions are now returned live, but the script output is still printed at the end (until we have better formatting).
+- Update integration tests to consume Startosis streaming endpoints
+
+# 0.53.12
+
+### Changes
+- Changed occurrences of `[sS]tartosis` to `Starlark` in errors sent by the CLI and its long and short description
+- Changed some logs and error messages inside core that which had references to Startosis to Starlark
+- Allow `dicts` & `structs` to be passed to `render_templates.config.data`
+
+# 0.53.11
+### Changes
+- Published the log-database HTTP port to the host machine
+
+# 0.53.10
+
+### Changes
+- Add 2 endpoints to the APIC that streams the output of a Startosis script execution
+- Changed the syntax of render_templates in Starlark
 
 ### Fixes
-- Better at handling errors when trying to format Kurtosis instructions at the CLI level
+- Fixed the error that would happen if there was a missing `kurtosis.mod` file at the root of the module
+
+# 0.53.9
+
+### Fixes
+- Renamed `artifact_uuid` to `artifact_id` and `src` to `src_path` in `upload_files` in Starlark
+
+# 0.53.8
 
 # 0.53.7
 ### Changes